Renamed BackupInitiator to ActionInitiator and moved it to separate class
